requirements.txt
python-digitalocean